requirements.txt
flask>=3.0.0
requests>=2.31.0
paho-mqtt>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.0.0
seaborn>=0.11.0